fix(mem/opt/sort): overflow, clique-exact, input-destroy, merge-bound
- MemoryAllocator.CreateList: floor min items at 1 when SizeOf>64 (was:
  64/size truncated to 0, losing the floor); skip ceilpow2 when
  newCapacity > 2^30 (was: ceilpow2 overflowed to negative for large
  capacity).
- UnsafeSlabAllocator.Allocated + AllocationCount: widen the size/count
  products to long (was: 32-bit overflow for many slabs / big T).
- MaximumClique.ShouldStop: drop the timer>1e8 abort (was: returned a
  suboptimal clique when the timer tripped, breaking the "exact"
  contract). Keep the valid solSize+candSize<=best bound prune.
- Rounding.Dependent: copy frac to a scratch buffer before mutating
  (was: zeroed the caller-owned frac[j] during sampling, destroying
  the input).
- MergeSorted.RunInPlaceCore: compute segment boundaries l+width /
  l+2*width / l+=2*width in long (was: 32-bit overflow for len>2^30
  where width reaches 2^30 and 2*width = 2^31 wraps negative).

Assessed false positive:
- GrayUnrank.Run: the `while (g > 0)` guard makes negative g return 0
  immediately (no infinite loop); positive g use identical
  arithmetic/logical shifts. Correct for valid non-negative input.

Deferred (perf-only, result correct):
- MeetInMiddle.Run: left table linear scan (slow, not wrong).
- Chromatic/Rook/Ntt: perf findings only, no correctness issues.

+++ src/IAFahim.Memory.Allocators/UnsafeSlabAllocator.cs
@@ -29,7 +29,7 @@ namespace IAFahim.Memory.Allocators
             *this.count = countPerSlab;
         }
 
-        public int AllocationCount => (this.countPerSlab * (this.slabs->Length - 1)) + *this.count;
+        public int AllocationCount => (int)((long)this.countPerSlab * (this.slabs->Length - 1)) + *this.count;
 
         public bool IsCreated => this.count != null;
 
@@ -70,7 +70,7 @@ namespace IAFahim.Memory.Allocators
 
         public int Allocated()
         {
-            return this.slabs->Length * this.countPerSlab * UnsafeUtility.SizeOf<T>();
+            return (int)((long)this.slabs->Length * this.countPerSlab * UnsafeUtility.SizeOf<T>());
         }
     }
 }
